requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# Deep Learning
tensorflow>=2.14.0,<2.16.0
keras>=2.14.0,<3.0.0
tf-keras-vis>=0.8.5

# Image Processing
pillow>=10.0.0
opencv-python-headless==4.9.0.80

# Data Science
numpy>=1.24.0,<2.0.0
pandas>=2.1.1
scikit-learn>=1.3.2

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0

# PDF Generation
reportlab>=4.0.0

# Authentication & Security
bcrypt>=4.0.1

# Explainability
lime>=0.2.0.1

# Utilities
pathlib>=1.0.1
